tests/wpt/tests/css/css-flexbox/parsing/webkit-aliases.html

<!DOCTYPE HTML>
<meta charset=UTF-8>
<title>CSS Flexbox Test: -webkit- legacy properties</title>
<link rel="help" href="https://www.w3.org/TR/css-flexbox-1/#webkit-aliases">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../support/shorthand-testcommon.js"></script>
<body>
<script>

test_is_legacy_name_alias("-webkit-align-content", "align-content");
test_is_legacy_name_alias("-webkit-align-items", "align-items");
test_is_legacy_name_alias("-webkit-align-self", "align-self");
test_is_legacy_name_alias("-webkit-flex", "flex");
test_is_legacy_name_alias("-webkit-flex-basis", "flex-basis");
test_is_legacy_name_alias("-webkit-flex-direction", "flex-direction");
test_is_legacy_name_alias("-webkit-flex-flow", "flex-flow");
test_is_legacy_name_alias("-webkit-flex-grow", "flex-grow");
test_is_legacy_name_alias("-webkit-flex-shrink", "flex-shrink");
test_is_legacy_name_alias("-webkit-flex-wrap", "flex-wrap");
test_is_legacy_name_alias("-webkit-justify-content", "justify-content");
test_is_legacy_name_alias("-webkit-order", "order");

</script>
